Add files via upload
The first file is a random password generator app where you can choose whether or not to include uppercase/lowercase characters, numbers, symbols, choose the length of the password, and generate a random password where you can also see it's strength. In the second file, I wanted to make a simple password generator in python, where the user answers with yes or no to questions about the uppercase/lowercase characters, numbers, and symbols.